Break up util.ts.
