+++ lsif/src/server/server.ts
@@ -15,7 +15,7 @@ import { createMetaRouter } from './routes/meta'
 import { createPostgresConnection } from '../shared/database/postgres'
 import { createQueue, ensureOnlyRepeatableJob } from '../shared/queue/queue'
 import { createTracer } from '../shared/tracing'
-import { dbFilename, dbFilenameOld, ensureDirectory } from '../shared/util'
+import { dbFilename, dbFilenameOld, ensureDirectory } from '../shared/paths'
 import { default as tracingMiddleware } from 'express-opentracing'
 import { defineRedisCommands } from './redis/redis'
 import { errorHandler } from './middleware/errors'
