+++ lsif/src/shared/settings.ts
@@ -1,0 +1,9 @@
+/**
+ * Reads an integer from an environment variable or defaults to the given value.
+ *
+ * @param key The environment variable name.
+ * @param defaultValue The default value.
+ */
+export function readEnvInt(key: string, defaultValue: number): number {
+    return (process.env[key] && parseInt(process.env[key] || '', 10)) || defaultValue
+}
