+++ lsif/src/shared/xrepo/bloom-filter.ts
@@ -1,5 +1,5 @@
 import { BloomFilter } from 'bloomfilter'
-import { readEnvInt } from '../util'
+import { readEnvInt } from '../settings'
 import { gzipJSON, gunzipJSON } from '../encoding/json'
 
 // These parameters give us a 1 in 1.38x10^9 false positive rate if we assume
